pyproject.toml
[project]
name = "transcript-scorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "google-adk>=1.18.0",
    "google-genai>=1.52.0",
    "language-tool-python>=3.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.6",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "vadersentiment>=3.3.2",
]
